Refactored radio-group and code cleanup.
- Refactored radio-group to be within radio-button folder to make application  tidy.
- Simplified IF statements.
- General code cleanup.
