+++ src/app/app.module.ts
@@ -27,7 +27,7 @@ import {DemoChipComponent} from './demo-chip/demo-chip.component';
 import { DemoMultiselectComponent } from './demo-multiselect/demo-multiselect.component';
 import { RadioButtonComponent } from './radio-button/radio-button.component';
 import { DemoRadioButtonComponent } from './demo-radio-button/demo-radio-button.component';
-import { RadioGroupComponent } from './radio-group/radio-group.component';
+import { RadioGroupComponent } from './radio-button/radio-group/radio-group.component';
 
 
 @NgModule({
