+++ src/app/demo-radio-button/demo-radio-button.component.html
@@ -1,7 +1,6 @@
 <div class="m-10">
   <div class="text-lg mb-4">Radio Buttons</div>
   <div class="border rounded p-6">
-
     <tw-radio-group [formControl]="formControl">
       <tw-radio-button label="One" value="ONE" class="block"></tw-radio-button>
       <tw-radio-button label="Two" value="TWO" class="block"></tw-radio-button>
