--- src/app/radio-group/radio.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { RadioService } from './radio.service';
-
-describe('RadioService', () => {
-  let service: RadioService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(RadioService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
